Optimise allocation code in Divine.Pool.

+++ divine/pool.ffi.h
@@ -3,8 +3,8 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-    void pool_extend( void *pool, size_t sz );
-    void pool_allocate( void *pool, size_t sz );
+    void pool_extend( void *pool, int sz );
+    void pool_allocate( void *pool, int sz );
 #ifdef __cplusplus
 }
 #endif
